requirements.txt
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-SocketIO==5.0.1
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mccabe==0.6.1
peewee==3.14.1
pyflakes==2.2.0
python-dotenv==0.19.0
python-engineio==4.0.0
python-socketio==5.0.4
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.3
websocket-client==0.57.0
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.2.0
